fixt(mhdflix): Refactor media response data classes for improved structure and organization
